requirements.txt
python-telegram-bot[job-queue]==21.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0